+++ example/src/main.rs
@@ -1,0 +1,4 @@
+use wasm_rs_dbg_example::example;
+fn main() {
+    example();
+}