+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- Support for `wasm32-wasi` target
+
 ## [0.1.0] - 2021-02-05
 
 Initial release